Move class into the ETL root
