+++ src/core/etl/tests/Flow/ETL/Tests/FlowTestCase.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Flow\ETL\Tests\Unit\Extractor;
+namespace Flow\ETL\Tests;
 
-use function Flow\ETL\DSL\rows;
 use Flow\ETL\{Config, Extractor, FlowContext, Rows};
 use PHPUnit\Framework\TestCase;
+use function Flow\ETL\DSL\rows;
 
 abstract class FlowTestCase extends TestCase
 {
