+++ src/core/etl/tests/Flow/ETL/Tests/Unit/Extractor/ChunkExtractorTest.php
@@ -6,6 +6,7 @@ namespace Flow\ETL\Tests\Unit\Extractor;
 
 use Flow\ETL\Extractor\ChunkExtractor;
 use Flow\ETL\Tests\Double\FakeExtractor;
+use Flow\ETL\Tests\FlowTestCase;
 
 final class ChunkExtractorTest extends FlowTestCase
 {
