+++ src/core/etl/tests/Flow/ETL/Tests/Unit/Extractor/DataFrameExtractorTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Flow\ETL\Tests\Unit\Extractor;
 
+use Flow\ETL\Tests\FlowTestCase;
 use function Flow\ETL\DSL\{df, from_data_frame, from_rows, row, rows, str_entry};
 
 final class DataFrameExtractorTest extends FlowTestCase
