+++ src/core/etl/tests/Flow/ETL/Tests/Unit/Extractor/MemoryExtractorTest.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Flow\ETL\Tests\Unit\Extractor;
 
-use function Flow\ETL\DSL\{from_memory, int_entry, str_entry, to_memory};
+use Flow\ETL\{Config, FlowContext, Row, Rows, Tests\FlowTestCase};
 use Flow\ETL\Memory\ArrayMemory;
-use Flow\ETL\{Config, FlowContext, Row, Rows};
+use function Flow\ETL\DSL\{from_memory, int_entry, str_entry, to_memory};
 
 final class MemoryExtractorTest extends FlowTestCase
 {
