+++ src/core/etl/tests/Flow/ETL/Tests/Unit/Extractor/PipelineExtractorTest.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace Flow\ETL\Tests\Unit\Extractor;
 
-use function Flow\ETL\DSL\int_entry;
+use Flow\ETL\{Row, Rows, Tests\FlowTestCase};
 use Flow\ETL\Extractor\{PipelineExtractor, RowsExtractor};
 use Flow\ETL\Pipeline\SynchronousPipeline;
-use Flow\ETL\{Row, Rows};
+use function Flow\ETL\DSL\int_entry;
 
 final class PipelineExtractorTest extends FlowTestCase
 {
