+++ src/core/etl/tests/Flow/ETL/Tests/Unit/Extractor/SequenceExtractorTest.php
@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 namespace Flow\ETL\Tests\Unit\Extractor;
 
+use Flow\ETL\{Row, Rows, Tests\FlowTestCase};
 use function Flow\ETL\DSL\{date_entry,
     float_entry,
     from_sequence_date_period,
     from_sequence_date_period_recurrences,
     from_sequence_number};
-use Flow\ETL\{Row, Rows};
 
 final class SequenceExtractorTest extends FlowTestCase
 {
